Partition InheritedRareData to reduce copy-on-write overhead from CSS zoom https://bugs.webkit.org/show_bug.cgi?id=313588 rdar://175800158
Reviewed by NOBODY (OOPS!).

When CSS zoom is applied to an element, every descendant that inherits usedZoom triggers
DataRef copy-on-write on InheritedRareData, copying all ~59 fields just to change one
float. This also breaks pointer sharing, so subsequent DataRef equality checks must do a
deep comparison instead of a pointer short-circuit.

This change partitions InheritedRareData into a small top-level struct containing the
frequently-mutated values (usedZoom, effectiveInert,  usedTouchAction, etc.) and a new
sub-DataRef containing the ~46 rarely-mutated CSS property fields. This follows the
existing NonInheritedData partitioning pattern.

When only usedZoom changes, the copy is now ~38 bytes (used values + one DataRef pointer)
instead of ~450 bytes. The CSS properties stay shared via the sub-DataRef. Style difference
checks also benefit from a pointer short-circuit on the cssData sub-DataRef.

For a 5,000 element document, this means that instead of using ~2.5 MB of additional
storage following a zoom operation, we only need ~190 KB.

Tests: perf/style-recalc-zoom.html

* LayoutTests/perf/style-recalc-zoom-expected.txt: Added.
* LayoutTests/perf/style-recalc-zoom.html: Added.
* Source/WebCore/Headers.cmake:
* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/css/CSSProperties.json:
* Source/WebCore/rendering/style/RenderStyle.cpp:
(WebCore::RenderStyle::copyPseudoElementsFrom):
* Source/WebCore/rendering/style/RenderStyle.h:
* Source/WebCore/rendering/updating/RenderTreeUpdater.cpp:
(WebCore::pseudoStyleCacheIsInvalid):
* Source/WebCore/style/StyleDifference.cpp:
* Source/WebCore/style/computed/StyleComputedStyle.cpp:
(WebCore::Style::ComputedStyle::inheritIgnoringCustomPropertiesFrom):
* Source/WebCore/style/computed/StyleComputedStyleBase+GettersInlines.h:
(WebCore::Style::ComputedStyleBase::inheritedCustomProperties const):
* Source/WebCore/style/computed/StyleComputedStyleBase.cpp:
(WebCore::Style::ComputedStyleBase::getCachedPseudoStyle const):
(WebCore::Style::ComputedStyleBase::addCachedPseudoStyle):
(WebCore::Style::ComputedStyleBase::setCustomPropertyValue):
(WebCore::Style::ComputedStyleBase::customPropertiesEqual const):
(WebCore::Style::ComputedStyleBase::deduplicateCustomProperties):
* Source/WebCore/style/computed/StyleComputedStyleBase.h:
(WebCore::Style::ComputedStyleBase::hasCachedPseudoStyles const):
* Source/WebCore/style/computed/StyleComputedStyleProperties+GettersCustomInlines.h:
(WebCore::Style::ComputedStyleProperties::cursor const):
* Source/WebCore/style/computed/StyleComputedStyleProperties+SettersCustomInlines.h:
(WebCore::Style::ComputedStyleProperties::setCursor):
* Source/WebCore/style/computed/data/StyleInheritedRareCSSData.cpp: Copied from Source/WebCore/style/computed/data/StyleInheritedRareData.cpp.
(WebCore::Style::InheritedRareCSSData::InheritedRareCSSData):
(WebCore::Style::InheritedRareCSSData::copy const):
(WebCore::Style::InheritedRareCSSData::operator== const):
(WebCore::Style::InheritedRareCSSData::dumpDifferences const):
* Source/WebCore/style/computed/data/StyleInheritedRareCSSData.h: Copied from Source/WebCore/style/computed/data/StyleInheritedRareData.h.
(WebCore::Style::InheritedRareCSSData::create):
* Source/WebCore/style/computed/data/StyleInheritedRareData.cpp:
(WebCore::Style::InheritedRareData::InheritedRareData):
(WebCore::Style::InheritedRareData::operator== const):
(WebCore::Style::InheritedRareData::dumpDifferences const): Ditto.
* Source/WebCore/style/computed/data/StyleInheritedRareData.h:
