+++ LayoutTests/perf/style-recalc-zoom.html
@@ -1,0 +1,71 @@
+<!DOCTYPE html>
+<html>
+<body>
+<script>
+if (window.testRunner)
+    testRunner.dumpAsText();
+
+var results = document.createElement('pre');
+document.body.appendChild(results);
+
+function log(msg) {
+    results.textContent += msg + '\n';
+}
+
+function createElements(parent, count) {
+    for (var i = 0; i < count; i++) {
+        var span = document.createElement('span');
+        parent.appendChild(span);
+    }
+}
+
+function measureCommittedBytes() {
+    if (window.GCController) {
+        GCController.collect();
+        GCController.collect();
+    }
+    if (window.internals)
+        return internals.mallocStatistics().committedVMBytes;
+    return 0;
+}
+
+var ELEMENT_COUNT = 5000;
+
+// Measure memory before creating anything.
+var before = measureCommittedBytes();
+
+// Create unzoomed elements and measure.
+var baselineContainer = document.createElement('div');
+document.body.appendChild(baselineContainer);
+createElements(baselineContainer, ELEMENT_COUNT);
+getComputedStyle(baselineContainer.lastElementChild).zoom;
+var midpoint = measureCommittedBytes();
+
+// Create zoomed elements and measure.
+var zoomedContainer = document.createElement('div');
+zoomedContainer.style.zoom = '2';
+document.body.appendChild(zoomedContainer);
+createElements(zoomedContainer, ELEMENT_COUNT);
+getComputedStyle(zoomedContainer.lastElementChild).zoom;
+var after = measureCommittedBytes();
+
+// Both containers exist simultaneously, so no memory was freed.
+var baselineCost = (midpoint - before) / ELEMENT_COUNT;
+var zoomedCost = (after - midpoint) / ELEMENT_COUNT;
+var perElementOverhead = zoomedCost - baselineCost;
+
+// With InheritedRareData partitioning, per-element zoom overhead
+// should be small (~38 bytes for the slimmed InheritedRareData copy,
+// with InheritedRareCSSData shared). Without partitioning, it would
+// be ~450 bytes per element. We use 200 bytes as the threshold.
+var THRESHOLD_BYTES_PER_ELEMENT = 200;
+
+log("Tests that per-element memory cost of zoomed elements vs unzoomed elements is below " + THRESHOLD_BYTES_PER_ELEMENT + " bytes.");
+
+if (perElementOverhead < THRESHOLD_BYTES_PER_ELEMENT)
+    log("PASS");
+else
+    log("FAIL: zoomed elements used " + Math.round(perElementOverhead) + " bytes/element more than unzoomed (threshold is " + THRESHOLD_BYTES_PER_ELEMENT + " bytes).");
+</script>
+</body>
+</html>
