+++ Source/WebCore/Headers.cmake
@@ -3138,6 +3138,7 @@ set(WebCore_PRIVATE_FRAMEWORK_HEADERS
     style/computed/data/StyleGridData.h
     style/computed/data/StyleGridItemData.h
     style/computed/data/StyleInheritedData.h
+    style/computed/data/StyleInheritedRareCSSData.h
     style/computed/data/StyleInheritedRareData.h
     style/computed/data/StyleMarqueeData.h
     style/computed/data/StyleMaskBorderData.h
